Updated code the USB print

+++ app/src/main/java/com/certify/snap/controller/PrinterController.java
@@ -486,10 +486,12 @@ public class PrinterController implements BCPControl.LIBBcpControlCallBack {
     }
 
     public void checkPrintFileExists() {
-        String path = Environment.getExternalStorageDirectory() + "/CertifySnap/Pic/image.jpg";
-        File file = new File(path);
-        if (file.exists()) {
-            file.delete();
+        if (AppSettings.isPrintUsbEnabled()) {
+            String path = Environment.getExternalStorageDirectory() + "/CertifySnap/Pic/image.jpg";
+            File file = new File(path);
+            if (file.exists()) {
+                file.delete();
+            }
         }
     }
